refactor(ruborist): move PeerDeps/DevDeps enums to model::node
Co-locate with EdgeType and NodeType for consistent module organization.
