+++ crates/ruborist/src/resolver/edges.rs
@@ -6,34 +6,9 @@ use petgraph::graph::{EdgeIndex, NodeIndex};
 
 use crate::model::graph::DependencyGraph;
 use crate::model::manifest::CoreVersionManifest;
-use crate::model::node::EdgeType;
+use crate::model::node::{DevDeps, EdgeType, PeerDeps};
 use crate::model::package_json::PackageJson;
 use crate::spec::{Catalogs, resolve_catalog_spec};
-
-/// Controls whether peer dependencies are included during edge iteration.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum PeerDeps {
-    /// Include peer dependencies (modern default).
-    Include,
-    /// Skip peer dependencies (legacy npm behaviour, `--legacy-peer-deps`).
-    Skip,
-}
-
-impl From<bool> for PeerDeps {
-    /// `true` (legacy_peer_deps) → `Skip`, `false` → `Include`.
-    fn from(legacy: bool) -> Self {
-        if legacy { Self::Skip } else { Self::Include }
-    }
-}
-
-/// Controls whether dev dependencies are included during edge iteration.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum DevDeps {
-    /// Include dev dependencies (root packages).
-    Include,
-    /// Exclude dev dependencies (transitive packages).
-    Exclude,
-}
 
 /// Represents an unresolved dependency edge extracted from the graph.
 #[derive(Debug, Clone)]
